refactor: domain 패키지 생성
- Card 관련 객체를 domain 으로 이동한다.
