+++ src/main/java/blackjack/domain/card/Card.java
@@ -1,4 +1,4 @@
-package balckjack.card;
+package blackjack.domain.card;
 
 public class Card {
     private final CardShape shape;
