+++ src/main/java/blackjack/domain/card/CardShape.java
@@ -1,4 +1,4 @@
-package balckjack.card;
+package blackjack.domain.card;
 
 public enum CardShape {
     DIAMOND(),
