+++ src/main/java/blackjack/domain/card/CardValue.java
@@ -1,4 +1,4 @@
-package balckjack.card;
+package blackjack.domain.card;
 
 public enum CardValue {
     TWO(),
